remove node modules

+++ Backend/controllers/sauceControllers.js
@@ -25,32 +25,37 @@ exports.create = (req, res, next) => {
 
 //Modification d'une sauce
 exports.update = (req, res, next) => {
-  const sauceObject = req.file ?
-    {
-      ...JSON.parse(req.body.sauce),
-      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-    } : { ...req.body };
-  if (req.file) {
-    Sauce.findOne({ _id: req.params.id })
-      .then((sauce) => {
-        const filename = sauce.imageUrl.split('/images/')[1];
-        fs.unlink(`images/${filename}`, () => {
-          Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-            .then(() => { res.status(200).json({ message: 'Sauce mise à jour!' }); })
-            .catch((error) => { res.status(400).json({ error }); });
+  if(sauce.userId != req.auth.userId){
+    res.status(401).json({message:"Utilisateur non autorisé !"})
+  } else{
+    const sauceObject = req.file ?
+      {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+      } : { ...req.body };
+    if (req.file) {
+      Sauce.findOne({ _id: req.params.id })
+        .then((sauce) => {
+          const filename = sauce.imageUrl.split('/images/')[1];
+          fs.unlink(`images/${filename}`, () => {
+            Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+              .then(() => { res.status(200).json({ message: 'Sauce mise à jour!' }); })
+              .catch((error) => { res.status(400).json({ error }); });
+          })
         })
-      })
-      .catch((error) => { res.status(500).json({ error }); });
+        .catch((error) => { res.status(500).json({ error }); });
 
-  } else {
-    Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Sauce mise à jour!' }))
-      .catch((error) => res.status(400).json({ error }));
+    } else {
+      Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+        .then(() => res.status(200).json({ message: 'Sauce mise à jour!' }))
+        .catch((error) => res.status(400).json({ error }));
+    }
   }
 };
 
 //Récupération de toutes les sauces
 exports.list = (req, res, next) => {
+  console.log("Liste des sauces")
   Sauce.find()
     .then((sauces) => {
       res.status(200).json(sauces);
@@ -81,12 +86,16 @@ exports.OneSauce = (req, res, next) => {
 exports.delete = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then(sauce => {
-      const filename = sauce.imageUrl.split('/images/')[1];
-      fs.unlink(`images/${filename}`, () => {
-        Sauce.deleteOne({ _id: req.params.id })
-          .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
-          .catch(error => res.status(400).json({ error }));
-      });
+      if(sauce.userId != req.auth.userId){
+        res.status(401).json({message:"Utilisateur non autorisé !"})
+      } else{
+        const filename = sauce.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
+            .catch(error => res.status(400).json({ error }));
+        });
+      }
     })
     .catch(error => res.status(500).json({ error }));
 };
